job_finder/requirements.txt
# Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM Providers
anthropic>=0.39.0
openai>=1.50.0

# Local LLM
ollama>=0.3.0

# Web framework
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0

# Database
chromadb>=0.5.0

# Browser automation
playwright>=1.48.0

# Embeddings & search
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# Document generation
python-docx>=1.1.0
reportlab>=4.2.0

# Email
imapclient>=3.0.0

# MCP
mcp>=1.0.0

# Utilities
python-dotenv>=1.0.0
cryptography>=43.0.0
httpx>=0.27.0
pyyaml>=6.0.0

# Resume parsing
pypdf>=4.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
